Adding scripts directory from Midway3
I have pulled the scripts directory from
midway, after organizing that directory
into the proper strucutre: data files in
the data directory, script files in the
scripts directory and I am trying to mirror
that structure locally. I think that I will
then be able to move files back and forth
with scp as need be.