Repo reorganized and network code added
I reorganized the repo in a coherent way
and added the code for fuzzy matching,
corpus filtering, network projecting. The
data files are not tracked and exist across
three locations: midway3, cronusx, and locally
on this machine. From now on use SCP to
transfer data, and do git pull to get the latest
code changes for wherever you are working.